+++ redwood-snapshot-testing/src/iosMain/kotlin/app/cash/redwood/snapshot/testing/UIViewSnapshotter.kt
@@ -20,18 +20,27 @@ import kotlin.time.DurationUnit
 import kotlinx.cinterop.useContents
 import platform.CoreGraphics.CGPointMake
 import platform.CoreGraphics.CGRectMake
+import platform.CoreGraphics.CGSizeMake
 import platform.UIKit.UIColor
 import platform.UIKit.UIScrollView
 import platform.UIKit.UIView
+import platform.UIKit.UIViewNoIntrinsicMetric
 
-/** Snapshot the subject on a white background. */
-class UIViewSnapshotter(
+/**
+ * Snapshot the subject on a white background.
+ *
+ * This either snapshots the view sized to be full-screen, or it uses [heightConstraint] to size the
+ * view to whatever height it requires. Scrolling snapshots are always sized to the screen height.
+ */
+class UIViewSnapshotter private constructor(
   private val callback: UIViewSnapshotCallback,
   private val subject: UIView,
+  private val widthConstraint: Constraint,
+  private val heightConstraint: Constraint,
 ) : Snapshotter {
 
   override fun snapshot(name: String?, scrolling: Boolean) {
-    layoutSubject()
+    layoutSubject(scrolling)
 
     // Unfortunately even with animations forced off, UITableView's animation system breaks
     // synchronous snapshots. The simplest workaround is to delay snapshots one frame.
@@ -73,26 +82,51 @@ class UIViewSnapshotter(
   }
 
   /** Do layout without taking a snapshot. */
-  fun layoutSubject() {
+  private fun layoutSubject(scrolling: Boolean) {
+    require(widthConstraint == Constraint.Fill) {
+      "width wrap not yet implemented"
+    }
+
+    if (heightConstraint == Constraint.Wrap && !scrolling) {
+      val widget = subject.subviews[0] as UIView
+
+      widget.setFrame(CGRectMake(0.0, 0.0, 0.0, 0.0))
+      val wrapSize = widget.sizeThatFits(
+        screenSize.useContents { CGSizeMake(width, UIViewNoIntrinsicMetric) },
+      )
+
+      val frame = wrapSize.useContents { CGRectMake(0.0, 0.0, width, height) }
+      subject.setFrame(frame)
+      widget.setFrame(frame)
+    }
+
     subject.layoutIfNeeded()
   }
 
   companion object {
+    private val screenSize = CGSizeMake(390.0, 844.0) // iPhone 14.
+    private val screenRect = screenSize.useContents { CGRectMake(0.0, 0.0, width, height) }
+
     fun framed(
       callback: UIViewSnapshotCallback,
       widget: UIView,
+      widthConstraint: Constraint = Constraint.Fill,
+      heightConstraint: Constraint = Constraint.Fill,
     ): UIViewSnapshotter {
       val frame = UIView()
         .apply {
-          val screenSize = CGRectMake(0.0, 0.0, 390.0, 844.0) // iPhone 14.
-
           backgroundColor = UIColor.whiteColor
-          setFrame(screenSize)
+          setFrame(screenRect)
 
-          widget.setFrame(screenSize)
+          widget.setFrame(screenRect)
           addSubview(widget)
         }
-      return UIViewSnapshotter(callback, frame)
+      return UIViewSnapshotter(callback, frame, widthConstraint, heightConstraint)
     }
+  }
+
+  enum class Constraint {
+    Fill,
+    Wrap,
   }
 }