+++ redwood-widget-uiview-test/RedwoodWidgetUIViewTests/UIViewRedwoodViewFillWrapTestHost.swift
@@ -1,0 +1,8 @@
+import RedwoodWidgetUIViewTestKt
+import UIKit
+
+final class UIViewRedwoodViewFillWrapTestHost: KotlinHostingXCTestCase<UIViewRedwoodViewFillWrapTest> {
+    override class func initTest(name: String) -> UIViewRedwoodViewFillWrapTest {
+        return UIViewRedwoodViewFillWrapTest(callback: SnapshotTestingCallback(named: name))
+    }
+}