+++ CHANGELOG.md
@@ -10,7 +10,8 @@ Changed:
 - Nothing yet!
 
 Fixed:
-- Nothing yet!
+- Don't measure the height of `RedwoodUIView` as zero when measured with `sizeThatFits()` or `intrinsicContentSize()`. We had been using `UIStackView` which doesn't support these functions.
+- Correctly signal `RedwoodUIView` size changes of to callers who measure it with `intrinsicContentSize()`.
 
 
 ## [0.18.0] - 2025-08-01